pyproject.toml
[project]
name = "simple-rag-with-openllmetry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.0.10",
    "llama-index>=0.12.37",
    "llama-index-vector-stores-chroma>=0.4.1",
    "openai==1.81.0",
    "traceloop-sdk>=0.40.7",
]
